requirements.txt
einops==0.8.0
matplotlib==3.10.0
ml_collections==1.0.0
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
scikit-image==0.25.0
scipy==1.14.1
six==1.17.0
tensorboard==2.18.0
torch==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
